refactor: Improve code quality
Refactor the Parser.parse method by moving all the sub functions to a different file as the method was getting too big.

Rename the AllCommand enum to AtlasCommand as these are commands related to the Atlas application.
